requirements.txt
numpy<1.24
scipy
librosa
torchmetrics
pytorch-lightning==1.9.5
pandas
matplotlib
scikit-learn
mpi4py
joblib==1.3.1
smplx
transforms3d
ftfy
axial_positional_embedding
einops
entmax
lightning-bolts
loguru
madgrad
mido
jsmin
optuna
Cython
git+https://github.com/CPJKU/madmom